+++ test/Main.hs
@@ -1,4 +1,8 @@
 module Main (main) where
 
+import qualified Spec as Spec
+import Test.Hspec
+
 main :: IO ()
-main = putStrLn "Test suite not yet implemented."
+main = hspec $ do
+  Spec.spec